Move CmdSignature and CmdEnvSuggestion to commandlib
